pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "procreate-to-tif"
version = "0.1.0"
description = "Convert .procreate archives into layered PSD, flat PNG/JPG, animated WebP/GIF, and stitched timelapse MP4."
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
authors = [
  { name = "Jerome Eno" },
]
requires-python = ">=3.9"
keywords = ["procreate", "psd", "image-conversion", "graphics", "macos"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: End Users/Desktop",
  "Intended Audience :: Developers",
  "Operating System :: MacOS",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]
dependencies = [
  "python-lzo",
  "Pillow",
  "numpy",
  "pytoshop",
  "six",
]

[project.optional-dependencies]
dev = [
  "build>=1.2",
  "pytest>=8.4",
]

[project.scripts]
procreate-to-tif = "procreate_to_tif.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
